Move cell to production
